+++ src/router/routes/index.js
@@ -1,3 +1,4 @@
+import assets from './assets'
 import mocabulary from './mocabulary'
 import mosys from './mosys'
 import piecemaker from './piecemaker'
@@ -5,6 +6,7 @@ import site from './site'
 import users from './users'
 
 export default [
+  assets,
   mocabulary,
   mosys,
   piecemaker,